+++ test/sandbox/test/rendering_test.rb
@@ -17,11 +17,11 @@ class RenderingTest < ViewComponent::TestCase
 
     allocations =
       if Rails.version.to_f < 8.0
-        {"3.3.8" => 124, "3.3.0" => 140, "3.2.8" => 122, "3.1.7" => 122, "3.0.7" => 131}
+        {"3.3" => 126, "3.2" => 128..129, "3.1" => 130, "3.0" => 131}
       elsif Rails.version.split(".").first(2).map(&:to_i) == [8, 0]
-        {"3.5.0" => 117, "3.4.4" => 121, "3.3.8" => 133}
+        {"3.5" => 118, "3.4" => 122, "3.3" => 134}
       else
-        {"3.4.4" => 119}
+        {"3.5" => 116, "3.4" => 120}
       end
 
     assert_allocations(**allocations) do